+++ src/cueTextTracks.js
@@ -7,6 +7,8 @@
 
 import videojs from 'video.js';
 
+const cueTextTracks = {};
+
 /**
 * This feature allows metadata text tracks to be manipulated once they are available,
 * usually after the 'loadstart' event is observed on the player
@@ -14,7 +16,7 @@ import videojs from 'video.js';
 * @param processMetadataTrack A callback that performs some operations on a
 * metadata text track
 **/
-export function processMetadataTracks(player, processMetadataTrack) {
+cueTextTracks.processMetadataTracks = function(player, processMetadataTrack) {
   const tracks = player.textTracks();
   const setModeAndProcess = function(track) {
     if (track.kind === 'metadata') {
@@ -24,24 +26,15 @@ export function processMetadataTracks(player, processMetadataTrack) {
   };
 
   // Text tracks are available
-  if (tracks.length > 0) {
-    for (let i = 0; i < tracks.length; i++) {
-      const track = tracks[i];
-
-      setModeAndProcess(track);
-    }
-  // Wait until text tracks are added
-  // We avoid always setting the event handler in case
-  // integrations decide to handle this separately
-  // with a different handler for the same event
-  } else {
-    tracks.addEventListener('addtrack', (event) => {
-      const track = event.track;
-
-      setModeAndProcess(track);
-    });
+  for (let i = 0; i < tracks.length; i++) {
+    setModeAndProcess(tracks[i]);
   }
-}
+
+  // Wait until text tracks are added
+  tracks.addEventListener('addtrack', (event) => {
+    setModeAndProcess(event.track);
+  });
+};
 
 /**
 * Sets the track mode to one of 'disabled', 'hidden' or 'showing'
@@ -49,20 +42,32 @@ export function processMetadataTracks(player, processMetadataTrack) {
 * Default behavior is to do nothing, @override if this is not desired
 * @param track The text track to set the mode on
 */
-export function setMetadataTrackMode(track) {
+cueTextTracks.setMetadataTrackMode = function(track) {
   return;
-}
+};
 
 /**
 * Determines whether cue is an ad cue and returns the cue data.
 * @param player A reference to the player
-* @param cue The cue to be checked
+* @param cue The full cue object
 * Returns the given cue by default @override if futher processing is required
-* @return the cueData in JSON if cue is a supported ad cue, or -1 if not
+* @return {Object} a useable ad cue or null if not supported
 **/
-export function getSupportedAdCue(player, cue) {
+cueTextTracks.getSupportedAdCue = function(player, cue) {
   return cue;
-}
+};
+
+/**
+* Defines whether a cue is supported or not, potentially
+* based on the player settings
+* @param player A reference to the player
+* @param cue The cue to be checked
+* Default behavior is to return true, @override if this is not desired
+* @return {Boolean}
+*/
+cueTextTracks.isSupportedAdCue = function(player, cue) {
+  return true;
+};
 
 /**
 * Gets the id associated with a cue.
@@ -70,9 +75,9 @@ export function getSupportedAdCue(player, cue) {
 * @returns The first occurance of 'id' in the object,
 * @override if this is not the desired cue id
 **/
-export function getCueId(player, cue) {
+cueTextTracks.getCueId = function(player, cue) {
   return cue.id;
-}
+};
 
 /**
 * Checks whether a cue has already been used
@@ -98,9 +103,9 @@ const setCueAlreadySeen = function(player, cueId) {
 * @param cues The set of cues to work with
 * @param processCue A method that uses a cue to make some
 * ad request in the ad implementation
-* @param [cancelAds] A method that dynamically cancels ads in the ad implementation
+* @param [cancelAdsHandler] A method that dynamically cancels ads in the ad implementation
 **/
-export function processAdTrack(player, cues, processCue, cancelAds) {
+cueTextTracks.processAdTrack = function(player, cues, processCue, cancelAdsHandler) {
   player.ads.includedCues = {};
 
   // loop over set of cues
@@ -109,7 +114,7 @@ export function processAdTrack(player, cues, processCue, cancelAds) {
     const cueData = this.getSupportedAdCue(player, cue);
 
     // Exit if this is not a supported cue
-    if (cueData === -1) {
+    if (!this.isSupportedAdCue(player, cue)) {
       videojs.log.warn('Skipping as this is not a supported ad cue.', cue);
       return;
     }
@@ -124,15 +129,18 @@ export function processAdTrack(player, cues, processCue, cancelAds) {
       return;
     }
 
+    // Optional dynamic ad cancellation
+    if (cancelAdsHandler) {
+      cancelAdsHandler(player, cueData, cueId, startTime);
+    }
+
     // Process cue as an ad cue
     processCue(player, cueData, cueId, startTime);
 
     // Indicate that this cue has been used
     setCueAlreadySeen(player, cueId);
 
-    // Optional dynamic ad cancellation
-    if (cancelAds !== undefined) {
-      cancelAds(player, cueData);
-    }
   }
-}
+};
+
+export default cueTextTracks;