+++ src/snapshot.js
@@ -15,7 +15,6 @@ import videojs from 'video.js';
  * @param {Object} player The videojs player object
  */
 export function getPlayerSnapshot(player) {
-
   let currentTime;
 
   if (videojs.browser.IS_IOS && player.ads.isLive(player)) {
@@ -79,7 +78,6 @@ export function getPlayerSnapshot(player) {
  * @param {Object} snapshotObject - the player state to apply
  */
 export function restorePlayerSnapshot(player, snapshotObject) {
-
   if (player.ads.disableNextSnapshotRestore === true) {
     player.ads.disableNextSnapshotRestore = false;
     return;
@@ -93,6 +91,7 @@ export function restorePlayerSnapshot(player, snapshotObject) {
 
   const suppressedRemoteTracks = snapshotObject.suppressedRemoteTracks;
   const suppressedTracks = snapshotObject.suppressedTracks;
+
   let trackSnapshot;
   const restoreTracks = function() {
     for (let i = 0; i < suppressedRemoteTracks.length; i++) {
@@ -200,8 +199,7 @@ export function restorePlayerSnapshot(player, snapshotObject) {
 
     // if the src changed for ad playback, reset it
     player.src({ src: snapshotObject.currentSrc, type: snapshotObject.type });
-    // safari requires a call to `load` to pick up a changed source
-    player.load();
+
     // and then resume from the snapshots time once the original src has loaded
     // in some browsers (firefox) `canplay` may not fire correctly.
     // Reace the `canplay` event with a timeout.